# Flat keyspace directory backend.