--- a/lib/backend/dir/doc.go
+++ b/lib/backend/dir/doc.go
@@ -12,16 +12,8 @@ distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
-
-fs package implements backend.Backend interface using a regular
-filesystem-based directory. The filesystem needs to be POSIX
-compliant and support 'date modified' attribute on files.
-
 */
 
-// Package 'dir' implements the "directory backend". It uses a regular
-// filesystem directories/files to store Teleport auth server state.
-//
-// Limitations:
-// 	- key names cannot start with '.' (dot)
+// dir package implements backend.Backend interface using the filesystem. The
+// filesystem needs to be POSIX compliant and support BSD locks (flock).
 package dir