--- a/lib/services/resource.go
+++ b/lib/services/resource.go
@@ -194,6 +194,16 @@ const (
 	VerbRotate = "rotate"
 )
 
+func CollectOptions(opts []MarshalOption) (*MarshalConfig, error) {
+	var cfg MarshalConfig
+	for _, o := range opts {
+		if err := o(&cfg); err != nil {
+			return nil, trace.Wrap(err)
+		}
+	}
+	return &cfg, nil
+}
+
 func collectOptions(opts []MarshalOption) (*MarshalConfig, error) {
 	var cfg MarshalConfig
 	for _, o := range opts {
@@ -208,6 +218,9 @@ func collectOptions(opts []MarshalOption) (*MarshalConfig, error) {
 type MarshalConfig struct {
 	// Version specifies particular version we should marshal resources with
 	Version string
+
+	// SkipValidation is used to skip schema validation.
+	SkipValidation bool
 }
 
 // GetVersion returns explicitly provided version or sets latest as default
@@ -231,6 +244,14 @@ func WithVersion(v string) MarshalOption {
 		default:
 			return trace.BadParameter("version '%v' is not supported", v)
 		}
+	}
+}
+
+// SkipValidation is used to disable schema validation.
+func SkipValidation() MarshalOption {
+	return func(c *MarshalConfig) error {
+		c.SkipValidation = true
+		return nil
 	}
 }