--- a/lib/services/server.go
+++ b/lib/services/server.go
@@ -10,6 +10,7 @@ import (
 	"github.com/gravitational/teleport/lib/utils"
 
 	"github.com/gravitational/trace"
+
 	"github.com/jonboulle/clockwork"
 )
 
@@ -270,7 +271,7 @@ const ServerSpecV2Schema = `{
     "cmd_labels": {
       "type": "object",
       "patternProperties": {
-        "^.*$": { 
+        "^.*$": {
           "type": "object",
           "additionalProperties": false,
           "required": ["command"],
@@ -431,19 +432,21 @@ func GetServerSchema() string {
 
 // UnmarshalServerResource unmarshals role from JSON or YAML,
 // sets defaults and checks the schema
-func UnmarshalServerResource(data []byte, kind string) (Server, error) {
+func UnmarshalServerResource(data []byte, kind string, cfg *MarshalConfig) (Server, error) {
 	if len(data) == 0 {
 		return nil, trace.BadParameter("missing server data")
 	}
+
 	var h ResourceHeader
-	err := json.Unmarshal(data, &h)
+	err := utils.FastUnmarshal(data, &h)
 	if err != nil {
 		return nil, trace.Wrap(err)
 	}
+
 	switch h.Version {
 	case "":
 		var s ServerV1
-		err := json.Unmarshal(data, &s)
+		err := utils.FastUnmarshal(data, &s)
 		if err != nil {
 			return nil, trace.Wrap(err)
 		}
@@ -451,8 +454,15 @@ func UnmarshalServerResource(data []byte, kind string) (Server, error) {
 		return s.V2(), nil
 	case V2:
 		var s ServerV2
-		if err := utils.UnmarshalWithSchema(GetServerSchema(), &s, data); err != nil {
-			return nil, trace.BadParameter(err.Error())
+
+		if cfg.SkipValidation {
+			if err := utils.FastUnmarshal(data, &s); err != nil {
+				return nil, trace.BadParameter(err.Error())
+			}
+		} else {
+			if err := utils.UnmarshalWithSchema(GetServerSchema(), &s, data); err != nil {
+				return nil, trace.BadParameter(err.Error())
+			}
 		}
 
 		if err := s.CheckAndSetDefaults(); err != nil {
@@ -481,20 +491,34 @@ func GetServerMarshaler() ServerMarshaler {
 // ServerMarshaler implements marshal/unmarshal of Role implementations
 // mostly adds support for extended versions
 type ServerMarshaler interface {
-	// UnmarshalServer from binary representation
-	UnmarshalServer(bytes []byte, kind string) (Server, error)
-	// MarshalServer to binary representation
+	// UnmarshalServer from binary representation.
+	UnmarshalServer(bytes []byte, kind string, opts ...MarshalOption) (Server, error)
+
+	// MarshalServer to binary representation.
 	MarshalServer(Server, ...MarshalOption) ([]byte, error)
+
+	// UnmarshalServers is used to unmarshal multiple servers from their
+	// binary representation.
+	UnmarshalServers(bytes []byte) ([]Server, error)
+
+	// MarshalServers is used to marshal multiple servers to their binary
+	// representation.
+	MarshalServers([]Server) ([]byte, error)
 }
 
 type TeleportServerMarshaler struct{}
 
 // UnmarshalServer unmarshals server from JSON
-func (*TeleportServerMarshaler) UnmarshalServer(bytes []byte, kind string) (Server, error) {
-	return UnmarshalServerResource(bytes, kind)
+func (*TeleportServerMarshaler) UnmarshalServer(bytes []byte, kind string, opts ...MarshalOption) (Server, error) {
+	cfg, err := collectOptions(opts)
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+
+	return UnmarshalServerResource(bytes, kind, cfg)
 }
 
-// MarshalServer marshals server into JSON
+// MarshalServer marshals server into JSON.
 func (*TeleportServerMarshaler) MarshalServer(s Server, opts ...MarshalOption) ([]byte, error) {
 	cfg, err := collectOptions(opts)
 	if err != nil {
@@ -503,7 +527,6 @@ func (*TeleportServerMarshaler) MarshalServer(s Server, opts ...MarshalOption) (
 	type serverv1 interface {
 		V1() *ServerV1
 	}
-
 	type serverv2 interface {
 		V2() *ServerV2
 	}
@@ -514,16 +537,44 @@ func (*TeleportServerMarshaler) MarshalServer(s Server, opts ...MarshalOption) (
 		if !ok {
 			return nil, trace.BadParameter("don't know how to marshal %v", V1)
 		}
-		return json.Marshal(v.V1())
+		return utils.FastMarshal(v.V1())
 	case V2:
 		v, ok := s.(serverv2)
 		if !ok {
 			return nil, trace.BadParameter("don't know how to marshal %v", V2)
 		}
-		return json.Marshal(v.V2())
+		return utils.FastMarshal(v.V2())
 	default:
 		return nil, trace.BadParameter("version %v is not supported", version)
 	}
+}
+
+// UnmarshalServers is used to unmarshal multiple servers from their
+// binary representation.
+func (*TeleportServerMarshaler) UnmarshalServers(bytes []byte) ([]Server, error) {
+	var servers []ServerV2
+
+	err := utils.FastUnmarshal(bytes, &servers)
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+
+	out := make([]Server, len(servers))
+	for i, v := range servers {
+		out[i] = Server(&v)
+	}
+	return out, nil
+}
+
+// MarshalServers is used to marshal multiple servers to their binary
+// representation.
+func (*TeleportServerMarshaler) MarshalServers(s []Server) ([]byte, error) {
+	bytes, err := utils.FastMarshal(s)
+	if err != nil {
+		return nil, trace.Wrap(err)
+	}
+
+	return bytes, nil
 }
 
 // SortedServers is a sort wrapper that sorts servers by name