--- a/lib/srv/regular/proxy.go
+++ b/lib/srv/regular/proxy.go
@@ -244,7 +244,7 @@ func (t *proxySubsys) proxyToHost(
 	// going to "local" CA? lets use the caching 'auth service' directly and avoid
 	// hitting the reverse tunnel link (it can be offline if the CA is down)
 	if site.GetName() == localDomain {
-		servers, err = t.srv.authService.GetNodes(t.namespace)
+		servers, err = t.srv.authService.GetNodes(t.namespace, services.SkipValidation())
 		if err != nil {
 			t.log.Warn(err)
 		}
@@ -254,7 +254,7 @@ func (t *proxySubsys) proxyToHost(
 		if err != nil {
 			t.log.Warn(err)
 		} else {
-			servers, err = siteClient.GetNodes(t.namespace)
+			servers, err = siteClient.GetNodes(t.namespace, services.SkipValidation())
 			if err != nil {
 				t.log.Warn(err)
 			}