**Purpose**

`tsh ssh` (and other commands like `tsh ls`) take a long time on large clusters. For example, on a 4k node cluster with a local etcd instance as the backend and auth, proxy, node all running in the same process, `tsh ssh` can take over 6 seconds. These numbers are larger for real world deployments where auth, proxy, and node run on different machines as well as etcd.

This is due to two reasons:

1. During login, `tsh` calls `GetNodes` several times. Each time `GetNodes` is called, it's validates the schema of the `services.Server` it returns. In the example 4k cluster, this in of itself takes up to 4 seconds. The solution to this is make schema validation optional in calls to `GetNodes`.

2. The remaining second is spent by the state cache re-upserting 4k items back into the dir backend. For example, writing 600 bytes to a new file takes about 200 us but writing 600 bytes to an open file takes about 10 us. To reduce the time spent re-upserting 4k items into the dir backend, the dir backend needs to support bulk insertion.

Fixes to these two problems should reduce login time to 500 ms to 1 second for 4k clusters.

**Implementation**

For the first problem, the API for `GetNodes(namespace string)` was updated to `GetNodes(namespace string, skipValidation bool)` allowing the caller to control if schema validation is run or not. In addition, `json-iterator` was used to speed up marshalling and unmarshalling.

For the second problem, the directory backend was re-written to support a flat keyspace. Once support for a flat keyspace was added, then the directory backend could store buckets with content being a list of keys. Then backend operations would simply be reading in all the keys in a bucket, operating on the map of items in the bucket, and then re-writing it back out. This reduces the time a upsert takes from 200 us to 10 us for each item, which for 4k nodes reduces total time from 800 ms to 40 ms.

**Related Issues**

Fixes https://github.com/gravitational/teleport/issues/2061