@russjones 2018-07-13
@kontsevoy Yes, it's in lib/backend/dir/migrate.go. It's relatively safe but there is a scenario where if the backend directories were poorly chosen it could fail.